Move `<CardBody />` to `card-body/` subfolder
