+++ packages/components/src/card/index.js
@@ -1,1 +1,2 @@
 export { default as Card } from './card';
+export { default as CardBody } from './card-body';
